Switch java client project to use Maven instead of Gradle (#98)
We use to publish new versions of the client on OSSRH, but that has been
turned down in favor of Maven Central Repository. And there's basically
only support for automated publishing via its Maven plugin - I couldn't
get it to work with Gradle. So this switches over all build, test,
signing, and publishing tasks to be configured with Maven, deleting the
Gradle counterparts. Also updates the GH Actions appropriately. Verified
that the [Publish
snapshot](https://github.com/tecton-ai/tecton-http-client-java/actions/runs/16436101362/job/46446300538)
action
[works](https://central.sonatype.com/repository/maven-snapshots/ai/tecton/java-client/0.9.4-SNAPSHOT/),
which shows that the new Maven setup can build, sign, and publish
appropriately. After we get some approval from customers (signifyd in
particular) that the published snapshot version works, we can run the
release and confirm that it works.